+++ featuretools/__init__.py
@@ -5,12 +5,14 @@ from . import variable_types
 from .entityset.api import *
 from . import primitives
 from .synthesis.api import *
-from .primitives import  list_primitives, install_primitives, Feature
+from .primitives import  list_primitives, install_primitives
 from .computational_backends.api import *
 from . import tests
 from .utils.pickle_utils import *
 from .utils.time_utils import *
 import featuretools.demo
 import featuretools.wrappers
+from . import feature_base
+from .feature_base import *
 
 __version__ = '0.5.1'
