+++ featuretools/feature_base/api.py
@@ -1,0 +1,7 @@
+from .feature_base import (
+    AggregationFeature,
+    DirectFeature,
+    FeatureBase,
+    IdentityFeature,
+    TransformFeature
+)
