+++ featuretools/primitives/base/api.py
@@ -3,10 +3,5 @@ from .aggregation_primitive_base import (
     AggregationPrimitive,
     make_agg_primitive
 )
-from .primitive_base import (
-    DirectFeature,
-    Feature,
-    IdentityFeature,
-    PrimitiveBase
-)
+from .primitive_base import PrimitiveBase
 from .transform_primitive_base import TransformPrimitive, make_trans_primitive
